added a small logo into the header
