+++ src/components/Feed/FeedComponent.css
@@ -7,12 +7,14 @@
   min-height: 1000px;
   min-width: 1500px;
   overflow: hidden;
-  border: none;
+  border: 2px solid black;
   border-radius: 1rem;
   background-image: url("../Images/GoT.jpg");
   background-size: cover;
   background-position: center;
   transition: 0.4 ease;
+  /* display: grid;
+  grid-template-columns: auto; */
 }
 
 @media screen and (min-width: 993px) and (max-width: 1600px) {
